+++ src/Thruster.cpp
@@ -1,25 +1,25 @@
 #include "Thruster.h"
 
-Eigen::Vector3f Thruster::viewThrust(){
+Eigen::Vector3d Thruster::viewThrust(){
 	return thrust;
 }
-Eigen::Vector3f Thruster::getThrust(){ //Need to add variance...
+Eigen::Vector3d Thruster::getThrust(){ //Need to add variance...
 	return thrust*power;
 } 
-void Thruster::setThrust(Eigen::Vector3f newThrust){
+void Thruster::setThrust(Eigen::Vector3d newThrust){
 	thrust = newThrust;
 }
 
-void Thruster::setPower(float newPower){
+void Thruster::setPower(double newPower){
 	power = newPower>1 ? 1 : (newPower<-1 ? -1 : newPower); //Constrains power to -1<power<1
 }
-float Thruster::getPower(){
+double Thruster::getPower(){
 	return power;
 }
 	
-void Thruster::setThrustVariance(Eigen::Vector3f newThrustVariance){
+void Thruster::setThrustVariance(Eigen::Vector3d newThrustVariance){
 	thrustVariance = newThrustVariance;
 }
-Eigen::Vector3f Thruster::getThrustVariance(){
+Eigen::Vector3d Thruster::getThrustVariance(){
 	return thrustVariance;
 }
